Order functions in modules
